Move out all logic regarding reading column to SqlTypesUnsafeHelper
